Refactor API endpoints into modular routers
