+++ app/main.py
@@ -3,12 +3,19 @@
 from contextlib import asynccontextmanager
 from pathlib import Path
 
-from api import endpoints
 from fastapi import FastAPI
 from fastapi.responses import HTMLResponse
 from fastapi.staticfiles import StaticFiles
 from utils import picks_by_name
 from utils.picks import store
+
+from app.api.routers import (
+	fbpick_router,
+	picks_router,
+	pipeline_router,
+	section_router,
+	upload_router,
+)
 
 
 @asynccontextmanager
@@ -32,7 +39,11 @@ app.mount(
 )
 
 # エンドポイント登録
-app.include_router(endpoints.router)
+app.include_router(upload_router)
+app.include_router(section_router)
+app.include_router(fbpick_router)
+app.include_router(pipeline_router)
+app.include_router(picks_router)
 
 
 @app.get('/', response_class=HTMLResponse)
