+++ scripts/__init__.py
@@ -1,0 +1,2 @@
+"""Utility scripts package."""
+
